fix golangci-lint

+++ internal/system/disk_unix.go
@@ -27,5 +27,5 @@ func DiskAvailable(path ...string) (uint64, error) {
 	if err := unix.Statfs(target, &stat); err != nil {
 		return 0, err
 	}
-	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
+	return stat.Bavail * uint64(stat.Bsize), nil
 }
